# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Materias API",
    "version": "1.0.0",
    "description": "API para la gestión de materias"
  },
  "servers": [
    {
      "url": "https://tutoriteso-fullversion.onrender.com/api"
    }
  ],
  "paths": {
    "/materias": {
      "get": {
        "summary": "Obtener todas las materias",
        "tags": ["Materias"],
        "responses": {
          "200": {
            "description": "Lista de materias obtenida exitosamente"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "post": {
        "summary": "Crear una nueva materia",
        "tags": ["Materias"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MateriaCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Materia creada exitosamente"
          },
          "400": {
            "description": "Petición inválida"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/materias/{id}": {
      "get": {
        "summary": "Obtener una materia por ID",
        "tags": ["Materias"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Materia encontrada exitosamente"
          },
          "404": {
            "description": "Materia no encontrada"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "put": {
        "summary": "Actualizar una materia",
        "tags": ["Materias"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MateriaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Materia actualizada exitosamente"
          },
          "404": {
            "description": "Materia no encontrada"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "delete": {
        "summary": "Eliminar una materia",
        "tags": ["Materias"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Materia eliminada exitosamente"
          },
          "404": {
            "description": "Materia no encontrada"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MateriaCreate": {
        "type": "object",
        "required": ["name", "description", "category", "url"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Matemáticas"
          },
          "description": {
            "type": "string",
            "example": "Curso de álgebra y cálculo"
          },
          "category": {
            "type": "string",
            "example": "Ciencias Exactas"
          },
          "url": {
            "type": "string",
            "example": "https://ejemplo.com/matematicas.jpg"
          }
        }
      },
      "MateriaUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Matemáticas II"
          },
          "description": {
            "type": "string",
            "example": "Curso de álgebra avanzada"
          },
          "category": {
            "type": "string",
            "example": "Ciencias Exactas"
          }
        }
      }
    }
  }
}